PGO: Added a 'Profile Guided Optimisations' target to the Makefile
The 'Profile Guided Optimisations' target allows you to build binaries
that are optimised for specific uses.

Here's a quick description on how to build a PGO binary :

$ make distclean
$ make pgo-firstpass
(Use the generated binary, on the same machine, like you're used to.
 This will generate gcda files which will be used during the final PGO
 build. gcda files can be manipulated using gcov)
$ make clean
$ make pgo-secondpass
(The binary is now optimised)

Informations about how to use this new build target were added to the
README.

This was added in response to :
https://community.arm.com/thread/10585

Signed-off-by: Myy <myy@miouyouyou.fr>